+++ bin/mega-evme/src/run/cmd.rs
@@ -137,7 +137,7 @@ impl Cmd {
         self.output_results(&outcome)?;
 
         // Step 5: Persist the RPC cache (clean-exit only).
-        cache_store.persist(None)?;
+        cache_store.persist()?;
 
         Ok(())
     }
